Add cost calculator
NB: We are aware that there is an Etherscan API key in this code.
Given that this is a free API key allowing 5 req/sec we will give
this a shot having it be public. If we run into issues here we'll
be forced to spin up a simple backend API with a CORS policy and
a cache for enabling this in the future.